Dialled it back a lot. It was all very exploratory so this commit is going to be a bit messy but I feel I'm at a point where I can start building the game up now.
I've had to remove the multiple rooms function in order to focus on just getting a simple card dealing mechanism going. The app can now:

 - Share a deck, burn pile and unique hands on the socket.io server.
 - Move cards from any collection of cards to another.

It might not sound like a lot but this is the basics of any card game. With this structure I can share and move cards around in an almost infinite number of ways.